CHANGE: Reorganized Functions
Created a utils file to contain all function for utility as opposed to
the pipeline file. The pipeline file will be used to process the data
and control the influx of data as it is fed for training the machine
learning model. The utils library will only contain the functions that
load the data and extract useful metadata from the dataset itself.

The description of the pyblight library was added to the init file.

A stats file was created for statistical analysis. This is where most of
the statistical functions will be created.
