+++ src/__init__.py
@@ -1,0 +1,9 @@
+"""Pyblight - A Helpful library for Healthcare-based ML.
+
+The pyblight package contains a series of functions and classes to
+process, analyze, and train health-based machine learning models. It
+provides classes for training machine learning models on the pytorch
+machine learning framework as well as functions that extract paths
+from image folders (with custom datasets that handle DICOM data), and
+functions that extracts and processes data from csv files.
+"""
