+++ src/stats.py
@@ -1,0 +1,45 @@
+"""Useful Functions for Statistical Analysis."""
+
+import torch
+
+def calculate_confusion_matrix(fin_predictions:pl.DataFrame):
+    """Calculate the confusion matrix using pandas.
+
+    Calculates the confusion matrix using a csv file that
+    contains both the predictions and actual labels. This
+    function then creates a crosstab of the data to develop
+    the confusion matrix.
+
+    Parameters
+    ----------
+    fin_predictions : Pandas DataFrame
+        DataFrame containing the prediction and actual
+        labels.
+
+    Returns
+    -------
+    Polars DataFrame
+        Cross tab containing the confusion matrix of the
+        predictions compared to the actual labels.
+
+    Dictionary
+        Contains the basic metrics obtained from the
+        confusion matrix. The metrics are the following:
+        - Accuracy
+        - Precision
+        - Recall
+        - F1 Score
+    """
+    ct = fin_predictions.pivot(values="predictions", index="classification", columns="classification", aggregate_function='count')
+    # Set the initial values
+    tp = ct.values[1][1]
+    tn = ct.values[0][0]
+    fn = ct.values[0][1]
+    fp = ct.values[1][0]
+    # Calculate the metrics
+    metrics = dict()
+    metrics['Accuracy'] = (tp + tn) / (tp + tn + fp + fn) # Ability of model to get the correct predictions
+    metrics['Precision'] = tp / (tp + fp) # Ability of model to label actual positives as positives (think retrospectively)
+    metrics['Recall'] = tp / (tp + fn) # Ability of model to correctly identify positives
+    metrics['F1 Score'] = (2 * metrics['Precision'] * metrics['Recall']) / (metrics['Precision'] + metrics['Recall'])
+    return ct, metrics
